Print the 98 first fibonaccis.

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,9 +6,9 @@
  */
 int main(void)
 {
-	int a = 1;
-	int b = 2;
-	int num;
+	long a = 1;
+	long b = 2;
+	long num;
 	int count = 2;
 
 	printf("%d, %d, ", a, b);
